myeongshim_rag/requirements.txt
fastapi
uvicorn
python-dotenv
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai
langchain-chroma
chromadb
pypdf
supabase
google-generativeai
